+++ packages/openai/src/image/openai-image-api.ts
@@ -1,7 +1,7 @@
 import { lazyValidator, zodSchema } from '@ai-sdk/provider-utils';
 import { z } from 'zod/v4';
 
-// minimal version of the schema, focussed on what is needed for the implementation
+// minimal version of the schema, focused on what is needed for the implementation
 // this approach limits breakages when the API changes and increases efficiency
 export const openaiImageResponseSchema = lazyValidator(() =>
   zodSchema(
@@ -9,7 +9,7 @@ export const openaiImageResponseSchema = lazyValidator(() =>
       data: z.array(
         z.object({
           b64_json: z.string(),
-          revised_prompt: z.string().optional(),
+          revised_prompt: z.string().nullish(),
         }),
       ),
     }),