+++ packages/amazon-bedrock/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @ai-sdk/amazon-bedrock
 
+## 3.0.51
+
+### Patch Changes
+
+- 51aa5de: backport: test server
+- Updated dependencies [51aa5de]
+  - @ai-sdk/provider-utils@3.0.16
+  - @ai-sdk/anthropic@2.0.41
+
+## 3.0.50
+
+### Patch Changes
+
+- Updated dependencies [f2da310]
+  - @ai-sdk/provider-utils@3.0.15
+  - @ai-sdk/anthropic@2.0.40
+
 ## 3.0.49
 
 ### Patch Changes