+++ packages/google-vertex/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @ai-sdk/google-vertex
 
+## 3.0.57
+
+### Patch Changes
+
+- 51aa5de: backport: test server
+- Updated dependencies [51aa5de]
+  - @ai-sdk/provider-utils@3.0.16
+  - @ai-sdk/anthropic@2.0.41
+  - @ai-sdk/google@2.0.27
+
+## 3.0.56
+
+### Patch Changes
+
+- Updated dependencies [f2da310]
+  - @ai-sdk/provider-utils@3.0.15
+  - @ai-sdk/anthropic@2.0.40
+  - @ai-sdk/google@2.0.26
+
 ## 3.0.55
 
 ### Patch Changes