+++ packages/huggingface/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @ai-sdk/huggingface
 
+## 0.0.8
+
+### Patch Changes
+
+- 51aa5de: backport: test server
+- Updated dependencies [51aa5de]
+  - @ai-sdk/openai-compatible@1.0.26
+  - @ai-sdk/provider-utils@3.0.16
+
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies [f2da310]
+  - @ai-sdk/provider-utils@3.0.15
+  - @ai-sdk/openai-compatible@1.0.25
+
 ## 0.0.6
 
 ### Patch Changes