+++ packages/rsc/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,40 @@
 
 ### Patch Changes
 
+- 51aa5de: backport: test server
+- Updated dependencies [51aa5de]
+  - ai@5.0.87
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [0f92046]
+  - ai@5.0.86
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@5.0.85
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [a017497]
+  - ai@5.0.84
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@5.0.83
+
+## 0.0.1
+
+### Patch Changes
+
 - ai@5.0.82
 
 ## 0.0.1