+++ packages/xai/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @ai-sdk/xai
 
+## 2.0.31
+
+### Patch Changes
+
+- 51aa5de: backport: test server
+- Updated dependencies [51aa5de]
+  - @ai-sdk/openai-compatible@1.0.26
+  - @ai-sdk/provider-utils@3.0.16
+
+## 2.0.30
+
+### Patch Changes
+
+- Updated dependencies [f2da310]
+  - @ai-sdk/provider-utils@3.0.15
+  - @ai-sdk/openai-compatible@1.0.25
+
 ## 2.0.29
 
 ### Patch Changes